+++ action_server/tests/conftest.py
@@ -1,7 +1,6 @@
 pytest_plugins = [
     "devutils.fixtures",
     "action_server_tests.fixtures",
-    "action_server_tests.data_server_cli_fixtures",
 ]